+++ appmap/test/data/pytest/test_simple.py
@@ -1,7 +1,21 @@
 import os
-
+import pytest
 
 def test_hello_world():
     import simple
     os.chdir('/tmp')
     assert simple.Simple().hello_world() == 'Hello world!'
+
+
+def test_status_failed():
+    assert False
+
+
+@pytest.mark.xfail
+def test_status_xfailed():
+    assert False
+
+
+@pytest.mark.xfail
+def test_status_xsucceeded():
+    assert True