+++ CHANGELOG.md
@@ -8,8 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - A recording is now created atomically. It is first written to a temp file, which is then
   renamed to the final file.
-  
 - Capture HTTP client request and response.
+
+### Fixed
+- When using `pytest` as the test driver, failed unittest cases appmaps are now recorded.
 
 ## [0.10.0] - 2021-05-07
 ### Added